Open gdb logfile after crash
Prior to this change deleting the logfile while the server was running
caused the logger to fail and the backtrace would be lost.

So now it starts logging and thus only creates the logfile when it
crashes which makes it much more stable. Also it does not create a big
file with thread start and stop lines.