Copy libantibot.so into bin folder as runtime dep
If there is a antibot folder configured.
It will look for a libantibot.so at the root of that folder.
And if it finds some it will copy it into the bin/ folder
next to the server binary.

If there is no libantibot.so found it will look for cpp
and h files to patch them into the current code base.

If there is no folder configured it will delete the libantibot.so