Smarter auto unlock build supporting CTRL-C
Catch all exists equally, success or any error case
including CTRL-C to unlock the build.

This way we do not have to manually catch all exits.
The lock should never get stuck because we do never build
twice in a row from one script.

Even ./loop_gdb.sh launches ./update.sh and as soon as it
is finished the build lock is freed.

If for whatever reason we need to build multiple times in
a row the manual `unlock_build` calls have
to be reintroduced.